Update from Original

+++ src/vs/workbench/parts/cli/electron-browser/cli.contribution.ts
@@ -5,7 +5,6 @@
 
 import * as nls from 'vs/nls';
 import * as path from 'path';
-import * as fs from 'fs';
 import * as os from 'os';
 import * as cp from 'child_process';
 import * as pfs from 'vs/base/node/pfs';
@@ -20,7 +19,7 @@ import { SyncActionDescriptor } from 'vs/platform/actions/common/actions';
 import { IMessageService, Severity } from 'vs/platform/message/common/message';
 import { IEditorService } from 'vs/platform/editor/common/editor';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import product from 'vs/platform/product';
+import product from 'vs/platform/node/product';
 
 interface ILegacyUse {
 	file: string;
@@ -37,8 +36,10 @@ function readOrEmpty(name: string): TPromise<string> {
 
 const root = URI.parse(require.toUrl('')).fsPath;
 const source = path.resolve(root, '..', 'bin', 'code');
-// TODO@Joao remove this, show the actions regardless and mention the actions can't run if that's the case
-const isAvailable = fs.existsSync(source);
+
+function isAvailable(): TPromise<boolean> {
+	return pfs.exists(source);
+}
 
 class InstallAction extends Action {
 
@@ -55,65 +56,73 @@ class InstallAction extends Action {
 	}
 
 	private get target(): string {
-		return `/usr/local/bin/${ product.applicationName }`;
+		return `/usr/local/bin/${product.applicationName}`;
 	}
 
 	run(): TPromise<void> {
-		return this.checkLegacy()
-			.then(uses => {
-				if (uses.length > 0) {
-					const { file, lineNumber } = uses[0];
-					const message = nls.localize(
-						'exists',
-						"Please remove the alias referencing '{0}' in '{1}' (line {2}) and retry this action.",
-						product.darwinBundleIdentifier,
-						file,
-						lineNumber
-					);
+		return isAvailable().then(isAvailable => {
+			if (!isAvailable) {
+				const message = nls.localize('not available', "This command is not available");
+				this.messageService.show(Severity.Info, message);
+				return undefined;
+			}
 
-					const resource = URI.file(file);
-					const input = { resource, mime: 'text/x-shellscript' };
-					const actions = [
-						new Action('inlineEdit', nls.localize('editFile', "Edit '{0}'", file), '', true, () => {
-							return this.editorService.openEditor(input).then(() => {
-								const message = nls.localize('again', "Please remove the '{0}' alias from '{1}' before continuing.", product.applicationName, file);
-								const actions = [
-									new Action('continue', nls.localize('continue', "Continue"), '', true, () => this.run()),
-									new Action('cancel', nls.localize('cancel', "Cancel"))
-								];
+			return this.checkLegacy()
+				.then(uses => {
+					if (uses.length > 0) {
+						const { file, lineNumber } = uses[0];
+						const message = nls.localize(
+							'exists',
+							"Please remove the alias referencing '{0}' in '{1}' (line {2}) and retry this action.",
+							product.darwinBundleIdentifier,
+							file,
+							lineNumber
+						);
 
-								this.messageService.show(Severity.Info, { message, actions });
-							});
+						const resource = URI.file(file);
+						const input = { resource };
+						const actions = [
+							new Action('inlineEdit', nls.localize('editFile', "Edit '{0}'", file), '', true, () => {
+								return this.editorService.openEditor(input).then(() => {
+									const message = nls.localize('again', "Please remove the '{0}' alias from '{1}' before continuing.", product.applicationName, file);
+									const actions = [
+										new Action('continue', nls.localize('continue', "Continue"), '', true, () => this.run()),
+										new Action('cancel', nls.localize('cancel', "Cancel"))
+									];
+
+									this.messageService.show(Severity.Info, { message, actions });
+								});
+							})
+						];
+
+						this.messageService.show(Severity.Warning, { message, actions });
+						return TPromise.as(null);
+					}
+
+					return this.isInstalled()
+						.then(isInstalled => {
+							if (!isAvailable || isInstalled) {
+								return TPromise.as(null);
+							} else {
+								const createSymlink = () => {
+									return pfs.unlink(this.target)
+										.then(null, ignore('ENOENT'))
+										.then(() => pfs.symlink(source, this.target));
+								};
+
+								return createSymlink().then(null, err => {
+									if (err.code === 'EACCES' || err.code === 'ENOENT') {
+										return this.createBinFolder()
+											.then(() => createSymlink());
+									}
+
+									return TPromise.wrapError(err);
+								});
+							}
 						})
-					];
-
-					this.messageService.show(Severity.Warning, { message, actions });
-					return TPromise.as(null);
-				}
-
-				return this.isInstalled()
-					.then(isInstalled => {
-						if (!isAvailable || isInstalled) {
-							return TPromise.as(null);
-						} else {
-							const createSymlink = () => {
-								return pfs.unlink(this.target)
-									.then(null, ignore('ENOENT'))
-									.then(() => pfs.symlink(source, this.target));
-							};
-
-							return createSymlink().then(null, err => {
-								if (err.code === 'EACCES' || err.code === 'ENOENT') {
-									return this.createBinFolder()
-										.then(() => createSymlink());
-								}
-
-								return TPromise.wrapError(err);
-							});
-						}
-					})
-					.then(() => this.messageService.show(Severity.Info, nls.localize('successIn', "Shell command '{0}' successfully installed in PATH.", product.applicationName)));
-			});
+						.then(() => this.messageService.show(Severity.Info, nls.localize('successIn', "Shell command '{0}' successfully installed in PATH.", product.applicationName)));
+				});
+		});
 	}
 
 	private isInstalled(): TPromise<boolean> {
@@ -185,13 +194,21 @@ class UninstallAction extends Action {
 	}
 
 	private get target(): string {
-		return `/usr/local/bin/${ product.applicationName }`;
+		return `/usr/local/bin/${product.applicationName}`;
 	}
 
 	run(): TPromise<void> {
-		return pfs.unlink(this.target)
-			.then(null, ignore('ENOENT'))
-			.then(() => this.messageService.show(Severity.Info, nls.localize('successFrom', "Shell command '{0}' successfully uninstalled from PATH.", product.applicationName)));
+		return isAvailable().then(isAvailable => {
+			if (!isAvailable) {
+				const message = nls.localize('not available', "This command is not available");
+				this.messageService.show(Severity.Info, message);
+				return undefined;
+			}
+
+			return pfs.unlink(this.target)
+				.then(null, ignore('ENOENT'))
+				.then(() => this.messageService.show(Severity.Info, nls.localize('successFrom', "Shell command '{0}' successfully uninstalled from PATH.", product.applicationName)));
+		});
 	}
 }
 
@@ -203,18 +220,24 @@ class DarwinCLIHelper implements IWorkbenchContribution {
 	) {
 		const installAction = instantiationService.createInstance(InstallAction, InstallAction.ID, InstallAction.LABEL);
 
-		installAction.checkLegacy().done(files => {
-			if (files.length > 0) {
-				const message = nls.localize('update', "Code needs to change the '{0}' shell command. Would you like to do this now?", product.applicationName);
-				const now = new Action('changeNow', nls.localize('changeNow', "Change Now"), '', true, () => installAction.run());
-				const later = new Action('later', nls.localize('later', "Later"), '', true, () => {
-					messageService.show(Severity.Info, nls.localize('laterInfo', "Remember you can always run the '{0}' action from the Command Palette.", installAction.label));
-					return null;
-				});
-				const actions = [now, later];
-
-				messageService.show(Severity.Info, { message, actions });
+		isAvailable().done(isAvailable => {
+			if (!isAvailable) {
+				return;
 			}
+
+			return installAction.checkLegacy().done(files => {
+				if (files.length > 0) {
+					const message = nls.localize('update', "Code needs to change the '{0}' shell command. Would you like to do this now?", product.applicationName);
+					const now = new Action('changeNow', nls.localize('changeNow', "Change Now"), '', true, () => installAction.run());
+					const later = new Action('later', nls.localize('later', "Later"), '', true, () => {
+						messageService.show(Severity.Info, nls.localize('laterInfo', "Remember you can always run the '{0}' action from the Command Palette.", installAction.label));
+						return null;
+					});
+					const actions = [now, later];
+
+					messageService.show(Severity.Info, { message, actions });
+				}
+			});
 		});
 	}
 
@@ -223,13 +246,13 @@ class DarwinCLIHelper implements IWorkbenchContribution {
 	}
 }
 
-if (isAvailable && process.platform === 'darwin') {
+if (process.platform === 'darwin') {
 	const category = nls.localize('shellCommand', "Shell Command");
 
-	const workbenchActionsRegistry = <IWorkbenchActionRegistry>Registry.as(ActionExtensions.WorkbenchActions);
+	const workbenchActionsRegistry = Registry.as<IWorkbenchActionRegistry>(ActionExtensions.WorkbenchActions);
 	workbenchActionsRegistry.registerWorkbenchAction(new SyncActionDescriptor(InstallAction, InstallAction.ID, InstallAction.LABEL), 'Shell Command: Install \'code\' command in PATH', category);
 	workbenchActionsRegistry.registerWorkbenchAction(new SyncActionDescriptor(UninstallAction, UninstallAction.ID, UninstallAction.LABEL), 'Shell Command: Uninstall \'code\' command from PATH', category);
 
-	const workbenchRegistry = <IWorkbenchContributionsRegistry>Registry.as(WorkbenchExtensions.Workbench);
+	const workbenchRegistry = Registry.as<IWorkbenchContributionsRegistry>(WorkbenchExtensions.Workbench);
 	workbenchRegistry.registerWorkbenchContribution(DarwinCLIHelper);
 }
